Move carbon footprint under the Energy view
The Carbon tab becomes a Carbon Footprint section at the bottom of the
Energy view: same totals, stacked daily chart and factor disclaimer,
rendered by the energy view itself. Removes the standalone route, nav
entry and carbon-view component.
