+++ src/views/energy-view.js
@@ -1,6 +1,7 @@
 import * as dataRepository from '../services/data-repository.js';
 import { DataView } from '../components/data-view/data-view.js';
 import { getChartColor, ChartColors } from '../utils/style.js';
+import { computeDailyFootprint, EMISSION_FACTORS, FLIGHT_DISTANCE_THRESHOLD_KM } from '../utils/carbon.js';
 
 export class EnergyView extends DataView {
     constructor() {
@@ -55,6 +56,59 @@ export class EnergyView extends DataView {
         await this.createSingleLineChart('gas-chart', 'Gas Import', 'gas_daily', 'usage_therms', getChartColor(ChartColors.Red), startDate, endDate);
 
         await this.createSingleLineChart('water-chart', 'Water Usage', 'water_daily', 'usage_liters', getChartColor(ChartColors.Blue), startDate, endDate);
+
+        this.renderCarbonSection(startDate, endDate);
+    }
+
+    /**
+     * Carbon footprint estimate derived from grid electricity, gas and
+     * travel distance (location history). See src/utils/carbon.js.
+     */
+    renderCarbonSection(startDate, endDate) {
+        const chartCard = this.querySelector('chart-card[chart-id="carbon-daily-chart"]');
+        if (!chartCard) return;
+
+        const electricityRows = dataRepository.getTimeSeriesData('electricity_grid_daily', startDate, endDate, 'ASC');
+        const gasRows = dataRepository.getTimeSeriesData('gas_daily', startDate, endDate, 'ASC');
+        const locationRows = dataRepository.getTimeSeriesData('location', startDate, endDate, 'ASC');
+
+        const daily = computeDailyFootprint({ electricityRows, gasRows, locationRows });
+
+        const sum = (key) => daily.reduce((acc, d) => acc + (d[key] || 0), 0);
+        const electricityTotal = sum('electricity_kg');
+        const gasTotal = sum('gas_kg');
+        const travelTotal = sum('travel_kg');
+        const total = electricityTotal + gasTotal + travelTotal;
+
+        const fmt = (kg) => kg >= 1000 ? `${(kg / 1000).toFixed(2)} t` : `${Math.round(kg)} kg`;
+        const setText = (id, value) => {
+            const el = this.querySelector(`#${id}`);
+            if (el) el.textContent = value;
+        };
+
+        setText('carbon-total', fmt(total));
+        setText('carbon-electricity', fmt(electricityTotal));
+        setText('carbon-gas', fmt(gasTotal));
+        setText('carbon-travel', fmt(travelTotal));
+        setText('carbon-disclaimer',
+            `Estimates only. Electricity: grid imports × ${EMISSION_FACTORS.electricity_kg_per_kwh} kg/kWh · ` +
+            `Gas: ${EMISSION_FACTORS.gas_kg_per_therm} kg/therm · ` +
+            `Travel: distance between location points × ${EMISSION_FACTORS.car_kg_per_km} kg/km ` +
+            `(${EMISSION_FACTORS.flight_kg_per_km} kg/km on days over ${FLIGHT_DISTANCE_THRESHOLD_KM} km, assumed flights).`);
+
+        chartCard.setDateRange(startDate, endDate);
+        chartCard.setTimeSeriesData(daily, {
+            series: [
+                { label: 'Electricity (kg)', key: 'electricity_kg', color: getChartColor(ChartColors.Yellow) },
+                { label: 'Gas (kg)', key: 'gas_kg', color: getChartColor(ChartColors.Red) },
+                { label: 'Travel (kg)', key: 'travel_kg', color: getChartColor(ChartColors.Blue) }
+            ],
+            startDate: startDate,
+            endDate: endDate,
+            interval: 'daily',
+            stacked: true,
+            fill: true
+        });
     }
 
     async createMultiLineChart(chartId, tableName, datasetsConfig, startDate, endDate) {
