+++ style.css
@@ -859,12 +859,19 @@ chart-card .chart-container {
     white-space: normal;
     word-wrap: break-word;
 }
-/* Carbon View */
-.carbon-container {
+/* Carbon Footprint section (Energy view) */
+.carbon-section {
     padding: 1rem;
     display: flex;
     flex-direction: column;
     gap: 1rem;
+}
+
+.carbon-section h2 {
+    font-size: 1rem;
+    text-transform: uppercase;
+    border-bottom: 2px solid var(--border-color);
+    padding-bottom: 0.5rem;
 }
 
 .carbon-disclaimer {
